Combined no sensor Excavation and Deposition into exdep
